+++ android/app/src/main/java/poke/rogue/helper/data/datasource/RemotePokemonListDataSource.kt
@@ -1,7 +1,7 @@
 package poke.rogue.helper.data.datasource
 
 import poke.rogue.helper.data.model.Pokemon
-import poke.rogue.helper.remote.dto.response.pokemon.toData
+import poke.rogue.helper.data.model.toData
 import poke.rogue.helper.remote.service.PokeDexService
 
 class RemotePokemonListDataSource(
