+++ android/app/src/test/java/poke/rogue/helper/remote/dto/response/pokemon/PokemonResponseTest.kt
@@ -4,6 +4,7 @@ import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
 import poke.rogue.helper.data.model.Pokemon
 import poke.rogue.helper.data.model.Type
+import poke.rogue.helper.data.model.toData
 import poke.rogue.helper.pokemonResponses
 import poke.rogue.helper.pokemonTypeResponses
